requirements.txt
numpy>=1.20.2
scipy>=1.6.2
psutil>=5.8.0
pyyaml>=6.0
matplotlib>=3.3.4
scikit-learn>=0.24.2
networkx>=2.6.3
seaborn>=0.11.1
colour>=0.1.5
pandas>=1.4.4
Cython >= 0.29.33